# transaction-service/src/main/resources/openapi.yaml
openapi: 3.0.1

info:
  title: Transaction API
  version: 1.0

paths:

  /api/transactions/deposit:

    post:

      tags:
        - Transaction

      operationId: depositAmount

      summary: Deposit money into account

      requestBody:
        required: true

        content:
          application/json:

            schema:
              $ref: '#/components/schemas/DepositRequest'

      responses:

        '200':
          description: Deposit successful

          content:
            application/json:

              schema:
                $ref: '#/components/schemas/TransactionResponse'

        '400':
          description: Invalid request

  /api/transactions/withdraw:

    post:

      tags:
        - Transaction

      operationId: withdrawAmount

      summary: Withdraw money from account

      requestBody:
        required: true

        content:
          application/json:

            schema:
              $ref: '#/components/schemas/WithdrawRequest'

      responses:

        '200':
          description: Withdrawal successful

          content:
            application/json:

              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/transactions/kyc/update:

    post:

      tags:
        - Transaction

      operationId: updateCustomerKyc

      summary: Update customer KYC

      requestBody:
        required: true

        content:
          application/json:

            schema:
              $ref: '#/components/schemas/UpdateKycRequest'

      responses:

        '200':
          description: KYC updated successfully

          content:
            application/json:

              schema:
                type: string
                example: KYC updated successfully

  /api/transactions/fund-transfer:

    post:

      tags:
        - Transaction

      operationId: fundTransfer

      summary: Transfer funds between accounts

      requestBody:
        required: true

        content:
          application/json:

            schema:
              $ref: '#/components/schemas/FundTransferRequest'

      responses:

        '200':
          description: Fund transfer successful

          content:
            application/json:

              schema:
                $ref: '#/components/schemas/FundTransferResponse'

  /api/transactions/process-interest:

    post:

      tags:
        - Transaction

      operationId: processDailyInterest

      summary: Process daily interest for all accounts

      responses:

        '200':
          description: Interest processed successfully

          content:
            application/json:

              schema:
                type: string
                example: Daily interest processed successfully

  /api/transactions/daily-report:

    get:

      tags:
        - Transaction

      operationId: getDailyTransactionReport

      summary: Get daily transaction report

      responses:

        '200':
          description: Daily report fetched successfully

          content:
            application/json:

              schema:
                $ref: '#/components/schemas/DailyTransactionReportResponse'

  /api/customers:

    post:

      tags:
        - Customer

      operationId: createCustomer

      summary: Create new customer

      requestBody:
        required: true

        content:
          application/json:

            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'

      responses:

        '200':
          description: Customer created successfully

          content:
            application/json:

              schema:
                $ref: '#/components/schemas/CreateCustomerResponse'


  /api/accounts:

    post:

      tags:
        - Account

      operationId: createAccount

      summary: Create account

      requestBody:
        required: true

        content:
          application/json:

            schema:
              $ref: '#/components/schemas/CreateAccountRequest'

      responses:

        '200':
          description: Account created successfully

          content:
            application/json:

              schema:
                $ref: '#/components/schemas/CreateAccountResponse'




components:

  schemas:

    DepositRequest:

      type: object

      required:
        - accountNumber
        - ifscCode
        - amount

      properties:

        accountNumber:
          type: string
          example: ACC1001
          minLength: 5
          maxLength: 20

        ifscCode:
          type: string
          minLength: 5
          example: SBIN0001234

        amount:
          type: string
          example: "786.28"

    WithdrawRequest:

      type: object

      required:
        - accountNumber
        - ifscCode
        - amount

      properties:

        accountNumber:
          type: string
          minLength: 5
          maxLength: 20
          example: ACC1001

        ifscCode:
          type: string
          minLength: 5
          example: SBIN0001234

        amount:
          type: string
          example: "786.28"

    TransactionResponse:

      type: object

      properties:

        message:
          type: string
          example: Amount deposited successfully

        accountNumber:
          type: string
          example: ACC1001

        updatedBalance:
          type: string
          example: "15000.24"

    CreateCustomerRequest:

      type: object

      required:
        - customerName
        - mobile
        - email
        - panNumber

      properties:

        customerName:
          type: string
          example: Abhishek

        mobile:
          type: string
          example: "9876543210"

        email:
          type: string
          example: test@gmail.com

        panNumber:
          type: string
          example: ABCDE1234F


    CreateCustomerResponse:

      type: object

      properties:

        customerId:
          type: integer
          format: int64
          example: 101

        message:
          type: string
          example: Customer created successfully


    CreateAccountRequest:

      type: object

      required:
        - holderName
        - email
        - accountType
        - initialBalance

      properties:

        holderName:
          type: string
          example: Abhishek Sharma

        email:
          type: string
          example: abhi@gmail.com

        accountType:
          type: string
          example: SAVINGS

        initialBalance:
          type: string
          example: "10000.00"



    CreateAccountResponse:

      type: object

      properties:

        accountNumber:
          type: string
          example: ACC1001

        ifscCode:
          type: string
          example: SBIN0001234

        message:
          type: string
          example: Account created successfully

    UpdateKycRequest:

      type: object

      required:
        - customerId
        - mobile
        - email

      properties:

        customerId:
          type: integer
          format: int64
          example: 101

        mobile:
          type: string
          example: 9876543210

        email:
          type: string
          example: test@gmail.com



    FundTransferRequest:

      type: object

      required:
        - fromAccount
        - toAccount
        - amount

      properties:

        fromAccount:
          type: string
          example: ACC1001

        toAccount:
          type: string
          example: ACC2002

        amount:
          type: string
          example: "786.28"

    FundTransferResponse:

      type: object

      properties:

        status:
          type: string
          example: SUCCESS

        message:
          type: string
          example: Fund transfer successful

        transactionId:
          type: integer
          format: int64
          example: 101


    DailyTransactionReportResponse:

      type: object

      properties:

        totalTransactions:
          type: integer
          example: 150

        totalDepositAmount:
          type: string
          example: "786.28"